Remove old ray_trace_3d view
