+++ raytracer/cherry-rs/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cherry-rs"
-version = "0.3.0"
+version = "1.0.0"
 authors = ["Kyle M. Douglass <kyle.m.douglass@gmail.com>"]
 repository = "https://github.com/kmdouglass/cherry"
 edition = "2021"
