+++ raytracer/cherry-rs/benches/convexplano_lens.rs
@@ -3,8 +3,8 @@ use std::hint::black_box;
 use std::rc::Rc;
 
 use cherry_rs::{
-    examples::convexplano_lens::sequential_model, n, ray_trace_3d_view, ray_trace_3d_view_v2,
-    ApertureSpec, FieldSpec, ParaxialView, PupilSampling, RefractiveIndexSpec,
+    examples::convexplano_lens::sequential_model, n, ray_trace_3d_view, ApertureSpec, FieldSpec,
+    ParaxialView, PupilSampling, RefractiveIndexSpec,
 };
 
 // Inputs
@@ -33,18 +33,6 @@ fn benchmark(c: &mut Criterion) {
     group.bench_function("ray_trace_3d_view", |b| {
         b.iter(|| {
             ray_trace_3d_view(
-                black_box(&APERTURE_SPEC),
-                black_box(&FIELD_SPECS),
-                black_box(&model),
-                black_box(&paraxial_view),
-                black_box(None),
-            )
-            .unwrap();
-        });
-    });
-    group.bench_function("ray_trace_3d_view_v2", |b| {
-        b.iter(|| {
-            ray_trace_3d_view_v2(
                 black_box(&APERTURE_SPEC),
                 black_box(&FIELD_SPECS),
                 black_box(&model),
