+++ raytracer/cherry-rs/src/views/ray_trace_3d/rays.rs
@@ -20,21 +20,25 @@ const TOL: Float = Float::EPSILON;
 pub struct Ray {
     pos: Vec3,
     dir: Vec3,
-    terminated: bool,
-    field_id: usize,
 }
 
 impl Ray {
-    pub fn new(pos: Vec3, dir: Vec3, field_id: usize) -> Result<Self> {
+    pub fn new(pos: Vec3, dir: Vec3) -> Result<Self> {
         if !dir.is_unit() {
             bail!("Ray direction must be a unit vector");
         }
-        Ok(Self {
-            pos,
-            dir,
-            terminated: false,
-            field_id,
-        })
+        Ok(Self { pos, dir })
+    }
+
+    /// Create a bundle of rays with default values.
+    pub fn new_bundle(num: usize) -> Vec<Self> {
+        vec![
+            Self {
+                pos: Vec3::new(0.0, 0.0, 0.0),
+                dir: Vec3::new(0.0, 0.0, 1.0),
+            };
+            num
+        ]
     }
 
     /// Finds the intersection point of a ray with a surface and the surface
@@ -137,13 +141,34 @@ impl Ray {
         self.dir = surf.rot_mat().transpose() * self.dir;
     }
 
-    pub fn terminate(&mut self) {
-        self.terminated = true;
+    // Return the x-coordinate of the ray position
+    pub fn x(&self) -> Float {
+        self.pos.x()
     }
 
-    #[inline]
-    pub fn is_terminated(&self) -> bool {
-        self.terminated
+    // Return the y-coordinate of the ray position
+    pub fn y(&self) -> Float {
+        self.pos.y()
+    }
+
+    // Return the z-coordinate of the ray position
+    pub fn z(&self) -> Float {
+        self.pos.z()
+    }
+
+    // Return the direction cosine k of the ray
+    pub fn k(&self) -> Float {
+        self.dir.k()
+    }
+
+    // Return the direction cosine l of the ray
+    pub fn l(&self) -> Float {
+        self.dir.l()
+    }
+
+    // Return the direction cosine m of the ray
+    pub fn m(&self) -> Float {
+        self.dir.m()
     }
 
     /// Create a fan of uniformly spaced rays in a given z-plane at an angle phi
@@ -172,7 +197,6 @@ impl Ray {
         phi: Float,
         radial_offset_x: Float,
         radial_offset_y: Float,
-        field_id: usize,
     ) -> Vec<Ray> {
         let pos = Vec3::fan(n, r, theta, z, radial_offset_x, radial_offset_y);
         let dir: Vec<Vec3> = pos
@@ -187,7 +211,7 @@ impl Ray {
 
         pos.iter()
             .zip(dir.iter())
-            .map(|(p, d)| Ray::new(*p, *d, field_id).unwrap())
+            .map(|(p, d)| Ray::new(*p, *d).unwrap())
             .collect()
     }
 
@@ -203,7 +227,6 @@ impl Ray {
     ///   amount in x
     /// - radial_offset_y: Offset the radial position of the vectors by this
     ///   amount in y
-    /// - `field_id`: Field ID of the rays.
     pub(crate) fn sq_grid_in_circ(
         radius: Float,
         spacing: Float,
@@ -211,7 +234,6 @@ impl Ray {
         phi: Float,
         radial_offset_x: Float,
         radial_offset_y: Float,
-        field_id: usize,
     ) -> Vec<Ray> {
         let theta = PI / 2.0; // TODO: For now rays are rotated about x only
 
@@ -229,7 +251,7 @@ impl Ray {
 
         pos.iter()
             .zip(dir.iter())
-            .map(|(p, d)| Ray::new(*p, *d, field_id).unwrap())
+            .map(|(p, d)| Ray::new(*p, *d).unwrap())
             .collect()
     }
 }
@@ -244,9 +266,8 @@ mod test {
     fn test_rays_new() {
         let pos = Vec3::new(0.0, 0.0, 0.0);
         let dir = Vec3::new(0.0, 0.0, 1.0);
-        let field_id = 0;
 
-        let rays = Ray::new(pos, dir, field_id);
+        let rays = Ray::new(pos, dir);
 
         assert!(rays.is_ok());
     }
@@ -256,9 +277,8 @@ mod test {
     fn test_rays_new_non_unit_dir() {
         let pos = Vec3::new(0.0, 0.0, 0.0);
         let dir = Vec3::new(0.0, 0.0, 2.0);
-        let field_id = 0;
 
-        let rays = Ray::new(pos, dir, field_id);
+        let rays = Ray::new(pos, dir);
 
         assert!(rays.is_err());
     }
@@ -266,14 +286,8 @@ mod test {
     // Test the intersection of a ray with a flat surface
     #[test]
     fn test_ray_intersection_flat_surface() {
-        let field_id = 0;
         let pos = Vec3::new(0.0, 0.0, 0.0);
-        let ray = Ray::new(
-            Vec3::new(0.0, 0.0, -1.0),
-            Vec3::new(0.0, 0.0, 1.0),
-            field_id,
-        )
-        .unwrap();
+        let ray = Ray::new(Vec3::new(0.0, 0.0, -1.0), Vec3::new(0.0, 0.0, 1.0)).unwrap();
         let surf_spec = SurfaceSpec::Conic {
             semi_diameter: 4.0,
             radius_of_curvature: Float::INFINITY,
@@ -291,13 +305,12 @@ mod test {
     // Test the intersection of a ray with a circular surface
     #[test]
     fn test_ray_intersection_conic() {
-        let field_id = 0;
         let pos = Vec3::new(0.0, 0.0, 0.0);
 
         // Ray starts at z = -1.0 and travels at 45 degrees to the optics axis
         let l = 0.7071;
         let m = ((1.0 as Float) - 0.7071 * 0.7071).sqrt();
-        let ray = Ray::new(Vec3::new(0.0, 0.0, -1.0), Vec3::new(0.0, l, m), field_id).unwrap();
+        let ray = Ray::new(Vec3::new(0.0, 0.0, -1.0), Vec3::new(0.0, l, m)).unwrap();
 
         // Surface has radius of curvature -1.0 and conic constant 0.0, i.e. a circle
         let surf_spec = SurfaceSpec::Conic {
